Added `date_taken_extractor`
Different media files all have different ways of encoding the timestamp
for when the media was originally recorded. This class is responsible
for being able to extract it.

If anything goes wrong, our strategy is to log it, and abort, leaving
the file in the staging area for further inspection.