Add rotation for video thumbnails
In my manual tests, I discovered that the `VideoStream` assumes the
largest direction is the width. The result is that the thumbnails for
vertical videos look sideways.

It looks like video players play the video in the right orientation
because there is a `Composite:Rotation` tag (on `mp4` videos at least)
that tells the number of degrees by which the video is rotated.

Now, the `ThumbnailGenerator` looks at the rotation tag, and rotates the
thumbnail to match the orientation of the original video.