Index relative paths
I realized that I didn't want absolute paths in the index, as that
would require either path munging on the client, or for the static file
server to run at `/`. Which is a terrible idea.

I decided to just index the relative paths. This also makes it easier
to move the haystack root location if necessary.