Disable indexing for video files
I thought it was ready, but apparently it isn't. :(

`ffvideo.VideoStream` segfaults when it tries to load videos on the
Raspberry Pi to generate a thumbnail. I'm not sure why yet, but I want
to be able to start indexing images while I figure out the video issues.

I opened an [issue][1] on the `ffvideo` project.

[1]: https://bitbucket.org/zakhar/ffvideo/issues/13/segfaults-on-raspberry-pi#