Use `exiftool` as context manager
Previously, I was having a problem where the `exiftool` processes
weren't being stopped when I shut down haystack. (Even with the `atexit`
registration)

Rather than keep trying to make that work, I decided to just refactor to
use the recommended context manager approach. This isn't great because
it will start a new `exiftool` process for each request, and then kill
it, resulting in a lot of churn, but I'm not too worried about that
since the indexer is going to be running constantly, and I don't care
too much how long it takes.