Disable indexer functionality
I've decided it's too much to try and take on all the problems that come
with managing media myself. (Like the timestamps, orientation, etc.) I
don't have time to handle all of that. However, it would still be nice
to get the automated transfer of media off my phone and onto a hard
drive. I believe haystack can still be used for this. This commit
disables most of the indexer functionality, reducing it to a simple
transfer mechanism.